+++ audio-player/assets/js/list-music.js
@@ -1,0 +1,38 @@
+export const listMusic = [
+  {
+    name: "Arsenium and Mianna and Heren",
+    author: "Bon Ami",
+    img: "./assets/images/track-covers/",
+    audio: "./assets/audio/Arsenium_and_Mianna_and_Heren_-_Bon_Ami.mp3"
+  },
+  {
+    name: "The PanHeads Band",
+    author: "Monster",
+    img: "./assets/images/track-covers/",
+    audio: "./assets/audio/The_PanHeads_Band-Monster.mp3"
+  },
+  {
+    name: "The PanHeads Band",
+    author: "SHjopot v temnote",
+    img: "./assets/images/track-covers",
+    audio: "./assets/audio/The_PanHeads_Band-SHjopot_v_temnote.mp3"
+  },
+  {
+    name: "ROCK_PRIVET",
+    author: "Nas ne dogonyat",
+    img: "./assets/images/track-covers/",
+    audio: "./assets/audio/ROCK_PRIVET_-_Nas_Ne_Dogonyat.mp3"
+  },
+  {
+    name: "The PanHeads Band",
+    author: "Я другой",
+    img: "./assets/images/track-covers/",
+    audio: "./assets/audio/The PanHeads Band — Я другой.mp3"
+  },
+  {
+    name: "The PanHeads Band",
+    author: "Poslannaya nebesami",
+    img: "./assets/images/track-covers/",
+    audio: "./assets/audio/The_PanHeads_Band_-_Poslannaya_nebesami.mp3"
+  }
+]
