+++ audio-player/index.html
@@ -15,7 +15,7 @@
         <progress max="100" value="20"></progress>
         <div class="time">0:00</div>
         <div class="time">0:00</div>
-        <audio src="./assets/audio/Arsenium_and_Mianna_and_Heren_-_Bon_Ami.mp3"></audio>
+        <!-- <audio src="./assets/audio/Arsenium_and_Mianna_and_Heren_-_Bon_Ami.mp3"></audio> -->
         <div class="player_wrapper_button">
           <button class="player_button_prev button">
             <img class="player_icon" src="./assets/images/svg/button-prev.svg" alt="button prev">
@@ -42,6 +42,6 @@
     </a>
     <span>&#169;2024</span>
   </footer>
-  <script src="./assets/index.js"></script> 
+  <script type="module" src="./assets/js/index.js"></script> 
 </body>
 </html>
